+++ e2e/e2etests/test_eth_deposit_liquidity_cap.go
@@ -6,11 +6,11 @@ import (
 	"cosmossdk.io/math"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/stretchr/testify/require"
+	"github.com/zeta-chain/protocol-contracts/pkg/gatewayevm.sol"
 
 	"github.com/zeta-chain/node/e2e/runner"
 	"github.com/zeta-chain/node/e2e/utils"
 	"github.com/zeta-chain/node/x/crosschain/types"
-	fungibletypes "github.com/zeta-chain/node/x/fungible/types"
 )
 
 // TestDepositEtherLiquidityCap tests depositing Ethers in a context where a liquidity cap is set
@@ -24,19 +24,17 @@ func TestDepositEtherLiquidityCap(r *runner.E2ERunner, args []string) {
 	liquidityCap := math.NewUintFromBigInt(supply).Add(liquidityCapArg)
 	amountLessThanCap := liquidityCapArg.BigInt().Div(liquidityCapArg.BigInt(), big.NewInt(10)) // 1/10 of the cap
 	amountMoreThanCap := liquidityCapArg.BigInt().Mul(liquidityCapArg.BigInt(), big.NewInt(10)) // 10 times the cap
-	msg := fungibletypes.NewMsgUpdateZRC20LiquidityCap(
-		r.ZetaTxServer.MustGetAccountAddressFromName(utils.OperationalPolicyName),
-		r.ETHZRC20Addr.Hex(),
-		liquidityCap,
-	)
-	res, err := r.ZetaTxServer.BroadcastTx(utils.OperationalPolicyName, msg)
+	res, err := r.ZetaTxServer.SetZRC20LiquidityCap(r.ETHZRC20Addr, liquidityCap)
 	require.NoError(r, err)
 
 	r.Logger.Info("set liquidity cap tx hash: %s", res.TxHash)
 	r.Logger.Info("Depositing more than liquidity cap should make cctx reverted")
 
-	signedTx, err := r.LegacySendEther(r.TSSAddress, amountMoreThanCap, nil)
-	require.NoError(r, err)
+	signedTx := r.ETHDeposit(
+		r.EVMAddress(),
+		amountMoreThanCap,
+		gatewayevm.RevertOptions{OnRevertGasLimit: big.NewInt(0)},
+	)
 
 	receipt := utils.MustWaitForTxReceipt(r.Ctx, r.EVMClient, signedTx, r.Logger, r.ReceiptTimeout)
 	utils.RequireTxSuccessful(r, receipt)
@@ -50,8 +48,11 @@ func TestDepositEtherLiquidityCap(r *runner.E2ERunner, args []string) {
 	initialBal, err := r.ETHZRC20.BalanceOf(&bind.CallOpts{}, r.EVMAddress())
 	require.NoError(r, err)
 
-	signedTx, err = r.LegacySendEther(r.TSSAddress, amountLessThanCap, nil)
-	require.NoError(r, err)
+	signedTx = r.ETHDeposit(
+		r.EVMAddress(),
+		amountLessThanCap,
+		gatewayevm.RevertOptions{OnRevertGasLimit: big.NewInt(0)},
+	)
 
 	receipt = utils.MustWaitForTxReceipt(r.Ctx, r.EVMClient, signedTx, r.Logger, r.ReceiptTimeout)
 	utils.RequireTxSuccessful(r, receipt)
@@ -68,22 +69,18 @@ func TestDepositEtherLiquidityCap(r *runner.E2ERunner, args []string) {
 	r.Logger.Info("Deposit succeeded")
 
 	r.Logger.Info("Removing the liquidity cap")
-	msg = fungibletypes.NewMsgUpdateZRC20LiquidityCap(
-		r.ZetaTxServer.MustGetAccountAddressFromName(utils.OperationalPolicyName),
-		r.ETHZRC20Addr.Hex(),
-		math.ZeroUint(),
-	)
-
-	res, err = r.ZetaTxServer.BroadcastTx(utils.OperationalPolicyName, msg)
+	res, err = r.ZetaTxServer.RemoveZRC20LiquidityCap(r.ETHZRC20Addr)
 	require.NoError(r, err)
-
 	r.Logger.Info("remove liquidity cap tx hash: %s", res.TxHash)
 
 	initialBal, err = r.ETHZRC20.BalanceOf(&bind.CallOpts{}, r.EVMAddress())
 	require.NoError(r, err)
 
-	signedTx, err = r.LegacySendEther(r.TSSAddress, amountMoreThanCap, nil)
-	require.NoError(r, err)
+	signedTx = r.ETHDeposit(
+		r.EVMAddress(),
+		amountMoreThanCap,
+		gatewayevm.RevertOptions{OnRevertGasLimit: big.NewInt(0)},
+	)
 
 	receipt = utils.MustWaitForTxReceipt(r.Ctx, r.EVMClient, signedTx, r.Logger, r.ReceiptTimeout)
 	utils.RequireTxSuccessful(r, receipt)