+++ x/emissions/types/params.go
@@ -10,12 +10,13 @@ import (
 // NewParams creates a new Params instance
 func NewParams() Params {
 	return Params{
-		ValidatorEmissionPercentage: "00.50",
-		ObserverEmissionPercentage:  "00.25",
-		TssSignerEmissionPercentage: "00.25",
-		ObserverSlashAmount:         ObserverSlashAmount,
-		BallotMaturityBlocks:        int64(BallotMaturityBlocks),
-		BlockRewardAmount:           BlockReward,
+		ValidatorEmissionPercentage:        "00.50",
+		ObserverEmissionPercentage:         "00.25",
+		TssSignerEmissionPercentage:        "00.25",
+		ObserverSlashAmount:                ObserverSlashAmount,
+		BallotMaturityBlocks:               int64(BallotMaturityBlocks),
+		BlockRewardAmount:                  BlockReward,
+		PendingBallotsDeletionBufferBlocks: PendingBallotsBufferBlocks,
 	}
 }
 
@@ -39,6 +40,9 @@ func (p Params) Validate() error {
 		return err
 	}
 	if err := validateBlockRewardsAmount(p.BlockRewardAmount); err != nil {
+		return err
+	}
+	if err := validatePendingBallotsBufferBlocks(p.PendingBallotsDeletionBufferBlocks); err != nil {
 		return err
 	}
 	return validateObserverSlashAmount(p.ObserverSlashAmount)
@@ -122,6 +126,18 @@ func validateBallotMaturityBlocks(i interface{}) error {
 		return fmt.Errorf("ballot maturity types must not be negative")
 	}
 
+	return nil
+}
+
+func validatePendingBallotsBufferBlocks(i interface{}) error {
+	v, ok := i.(int64)
+	if !ok {
+		return fmt.Errorf("invalid parameter type: %T", i)
+	}
+
+	if v < 1 {
+		return fmt.Errorf("pending ballots buffer blocks must not be less that 1")
+	}
 	return nil
 }